requirements.txt
altair==5.5.0
annotated-types==0.6.0
anyio==4.11.0
aom==3.6.0
arrow-cpp==21.0.0
asttokens==3.0.0
attrs==25.4.0
aws-c-auth==0.9.0
aws-c-cal==0.9.2
aws-c-common==0.12.4
aws-c-compression==0.3.1
aws-c-event-stream==0.5.6
aws-c-http==0.10.4
aws-c-io==0.21.4
aws-c-mqtt==0.13.3
aws-c-s3==0.8.7
aws-c-sdkutils==0.2.4
aws-checksums==0.2.7
aws-crt-cpp==0.34.0
aws-sdk-cpp==1.11.638
blas==1.0
blinker==1.9.0
bottleneck==1.4.2
brotlicffi==1.0.9.2
bzip2==1.0.8
c-ares==1.34.5
cachetools==5.5.1
cairo==1.18.4
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.2.1
colorama==0.4.6
comm==0.2.3
dav1d==1.2.1
debugpy==1.8.17
decorator==5.2.1
distro==1.9.0
exceptiongroup==1.3.0
executing==2.2.1
expat==2.7.3
fontconfig==2.15.0
freeglut==3.4.0
freetype==2.13.3
fribidi==1.0.10
gflags==2.2.2
gitdb==4.0.12
gitpython==3.1.45
glog==0.5.0
graphite2==1.3.14
h11==0.16.0
h2==4.3.0
harfbuzz==10.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
icu==73.1
idna==3.11
importlib-metadata==8.7.0
intel-openmp==2025.0.0
ipykernel==7.1.0
ipython==9.7.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jinja2==3.1.6
jiter==0.6.1
jpeg==9f
jsonschema==4.25.0
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.9.1
krb5==1.21.3
lcms2==2.16
lerc==4.0.0
libabseil==20250127.0
libavif==1.1.1
libbrotlicommon==1.0.9
libbrotlidec==1.0.9
libbrotlienc==1.0.9
libcurl==8.16.0
libdeflate==1.22
libglib==2.84.4
libgrpc==1.71.0
libiconv==1.16
libkrb5==1.21.3
libopenjpeg==2.5.4
libpng==1.6.50
libprotobuf==5.29.3
libre2-11==2024.07.02
libsodium==1.0.20
libssh2==1.11.1
libthrift==0.15.0
libtiff==4.7.0
libwebp-base==1.3.2
libxml2==2.13.9
libzlib==1.3.1
lz4-c==1.9.4
markupsafe==3.0.2
matplotlib-inline==0.2.1
mkl==2025.0.0
mkl-service==2.5.2
mkl_fft==2.1.1
mkl_random==1.3.0
narwhals==2.7.0
nest-asyncio==1.6.0
numexpr==2.14.1
numpy==2.3.4
numpy-base==2.3.4
openai==1.109.1
orc==2.2.0
packaging==25.0
pandas==2.3.3
parso==0.8.5
pcre2==10.46
pillow==12.0.0
pip==25.2
pixman==0.46.4
platformdirs==4.5.0
prompt-toolkit==3.0.52
protobuf==5.29.3
psutil==7.1.3
pure_eval==0.2.3
pyarrow==21.0.0
pycparser==2.23
pydantic==2.12.3
pydantic-core==2.41.4
pygments==2.19.2
pysocks==1.7.1
python==3.11.14
python-dateutil==2.9.0post0
python-dotenv==1.2.1
python-tzdata==2025.2
python_abi==3.11
pytz==2025.2
pywin32==311
pyyaml==6.0.2
pyzmq==27.1.0
re2==2024.07.02
referencing==0.37.0
requests==2.32.5
rpds-py==0.28.0
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
snappy==1.2.1
sniffio==1.3.1
sqlite==3.51.0
stack_data==0.6.3
streamlit==1.51.0
tbb==2022.0.0
tbb-devel==2022.0.0
tenacity==9.1.2
tk==8.6.15
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typing-extensions==4.15.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025b
ucrt==10.0.22621.0
urllib3==2.5.0
utf8proc==2.6.1
vc==14.42
vc14_runtime==14.44.35208
vs2015_runtime==14.44.35208
watchdog==6.0.0
wcwidth==0.2.14
wheel==0.45.1
win_inet_pton==1.1.0
xz==5.6.4
yaml==0.2.5
zeromq==4.3.5
zipp==3.23.0
zlib==1.3.1
zstd==1.5.7
